+++ src/app/data_harvester/disks/unix/linux/mod.rs
@@ -1,5 +1,5 @@
 mod partition;
 pub(crate) use partition::*;
 
-mod io_counters;
-pub use io_counters::*;
+mod counters;
+pub use counters::*;
