+++ src/app/data_harvester/disks/unix/macos/counters.rs
@@ -3,27 +3,7 @@
 use std::ffi::OsStr;
 
 use super::io_kit::{self, get_dict, get_disks, get_i64, get_string};
-
-#[derive(Debug, Default)]
-pub struct IoCounters {
-    name: String,
-    read_bytes: u64,
-    write_bytes: u64,
-}
-
-impl IoCounters {
-    pub(crate) fn device_name(&self) -> &OsStr {
-        OsStr::new(&self.name)
-    }
-
-    pub(crate) fn read_bytes(&self) -> u64 {
-        self.read_bytes
-    }
-
-    pub(crate) fn write_bytes(&self) -> u64 {
-        self.write_bytes
-    }
-}
+use crate::app::data_harvester::disks::IoCounters;
 
 fn get_device_io(device: io_kit::IoObject) -> anyhow::Result<IoCounters> {
     let parent = device.service_parent()?;
